requirements.txt
gradio
torch
transformers
Pillow
scipy
requests
selenium
